+++ ironfish/src/wallet/exporter/accountImport.ts
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { Account } from '../account/account'
-import { MultisigKeysImport } from '../interfaces/multisigKeys'
 import { HeadValue } from '../walletdb/headValue'
+import { MultisigKeysImport } from './multisig'
 
 export type AccountImport = {
   version: number
