+++ ironfish/src/wallet/exporter/multisig.ts
@@ -2,6 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import { multisig } from '@ironfish/rust-nodejs'
+import { MultisigKeys, MultisigSigner } from '../interfaces/multisigKeys'
 import {
   AccountDecodingOptions,
   MultisigIdentityEncryption,
@@ -36,4 +37,25 @@ export function decodeEncryptedMultisigAccount(
   } catch (e) {
     throw new Error(`Failed to decrypt multisig account: ${String(e)}`)
   }
+}
+
+// Multisig signing data can come from:
+// 1. Regular account export and imported which will have the secret
+// 2. Import from a trusted dealer, which will only have the identity
+export type MultisigKeysImport = MultisigKeys | MultisigSignerTrustedDealerImport
+
+export interface MultisigSignerTrustedDealerImport {
+  identity: string
+  keyPackage: string
+  publicKeyPackage: string
+}
+
+export function isMultisigSignerImport(data: MultisigKeysImport): data is MultisigSigner {
+  return 'secret' in data
+}
+
+export function isMultisigSignerTrustedDealerImport(
+  data: MultisigKeysImport,
+): data is MultisigSignerTrustedDealerImport {
+  return 'identity' in data
 }
