+++ ironfish/src/wallet/interfaces/multisigKeys.ts
@@ -7,29 +7,8 @@ export interface MultisigSigner {
   publicKeyPackage: string
 }
 
-export interface MultisigSignerTrustedDealerImport {
-  identity: string
-  keyPackage: string
-  publicKeyPackage: string
-}
-
 export interface MultisigCoordinator {
   publicKeyPackage: string
 }
 
 export type MultisigKeys = MultisigSigner | MultisigCoordinator
-
-// Multisig signing data can come from:
-// 1. Regular account export and imported which will have the secret
-// 2. Import from a trusted dealer, which will only have the identity
-export type MultisigKeysImport = MultisigKeys | MultisigSignerTrustedDealerImport
-
-export function isMultisigSignerImport(data: MultisigKeysImport): data is MultisigSigner {
-  return 'secret' in data
-}
-
-export function isMultisigSignerTrustedDealerImport(
-  data: MultisigKeysImport,
-): data is MultisigSignerTrustedDealerImport {
-  return 'identity' in data
-}
